requirements.txt
py-cord==2.4.1
requests==2.31.0
python-dotenv==1.0.1
aiohttp==3.8.6